Move parsing items to separate namespace
