+++ Source/MochaTool.InteropGen/CodeGen/ManagedCodeGenerator.cs
@@ -1,4 +1,5 @@
-﻿using System.CodeDom.Compiler;
+﻿using MochaTool.InteropGen.Parsing;
+using System.CodeDom.Compiler;
 using System.Collections.Immutable;
 
 namespace MochaTool.InteropGen.CodeGen;
