+++ Source/MochaTool.InteropGen/Program.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using MochaTool.InteropGen.CodeGen;
 using MochaTool.InteropGen.Extensions;
+using MochaTool.InteropGen.Parsing;
 
 namespace MochaTool.InteropGen;
